+++ Runtime/Purchasing/Analytics/EmptyUnityAnalyticsAdapter.cs
@@ -1,15 +1,9 @@
-using Unity.Services.Analytics;
-
 namespace UnityEngine.Purchasing
 {
     class EmptyAnalyticsAdapter : IAnalyticsAdapter
     {
-        public void SendTransactionEvent(Product product)
-        {
-        }
+        public void SendTransactionEvent(Product product) { }
 
-        public void SendTransactionFailedEvent(Product product, PurchaseFailureReason reason)
-        {
-        }
+        public void SendTransactionFailedEvent(Product product, PurchaseFailureReason reason) { }
     }
 }